+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Viewer: optionally show rating & tags on overlay
 - Viewer: long press on rating quick action for quicker rating
 - Search: missing address filter
 - Lithuanian translation (thanks Gediminas Murauskas)